// dolphinscheduler-ui/src/views/projects/trigger/components/node/types.ts
interface ITaskPriorityOption extends SelectOption {
  icon: VNode
  color: string
}